Casts to integer inf

+++ src/poli/benchmarks/pmo.py
@@ -75,7 +75,7 @@ class PMOBenchmark(GuacaMolGoalDirectedBenchmark):
         batch_size: Union[int, None] = None,
         parallelize: bool = False,
         num_workers: Union[int, None] = None,
-        evaluation_budget: int = float("inf"),
+        evaluation_budget: int = int("inf"),
     ) -> None:
         super().__init__(
             string_representation=string_representation,
