+++ src/poli/objective_repository/albuterol_similarity/register.py
@@ -71,7 +71,7 @@ class AlbuterolSimilarityBlackBox(TDCBlackBox):
 
     Methods
     -------
-    __init__(self, string_representation, force_isolation, batch_size=None, parallelize=False, num_workers=None, evaluation_budget=float("inf"))
+    __init__(self, string_representation, force_isolation, batch_size=None, parallelize=False, num_workers=None, evaluation_budget=int("inf"))
         Initializes an instance of the black box.
 
     References
@@ -93,7 +93,7 @@ class AlbuterolSimilarityBlackBox(TDCBlackBox):
         batch_size: int = None,
         parallelize: bool = False,
         num_workers: int = None,
-        evaluation_budget: int = float("inf"),
+        evaluation_budget: int = int("inf"),
     ):
         super().__init__(
             oracle_name="Albuterol_Similarity",
@@ -152,7 +152,7 @@ class AlbuterolSimilarityProblemFactory(AbstractProblemFactory):
         batch_size: int = None,
         parallelize: bool = False,
         num_workers: int = None,
-        evaluation_budget: int = float("inf"),
+        evaluation_budget: int = int("inf"),
         force_isolation: bool = False,
     ) -> Problem:
         """
