+++ src/poli/objective_repository/zaleplon_mpo/register.py
@@ -87,7 +87,7 @@ class ZaleplonMPOBlackBox(TDCBlackBox):
         batch_size: int = None,
         parallelize: bool = False,
         num_workers: int = None,
-        evaluation_budget: int = float("inf"),
+        evaluation_budget: int = int("inf"),
     ):
         super().__init__(
             oracle_name="Zaleplon_MPO",
@@ -146,7 +146,7 @@ class ZaleplonMPOProblemFactory(AbstractProblemFactory):
         batch_size: int = None,
         parallelize: bool = False,
         num_workers: int = None,
-        evaluation_budget: int = float("inf"),
+        evaluation_budget: int = int("inf"),
         force_isolation: bool = False,
     ) -> Problem:
         """
